requirements.txt
openai-whisper
openai
fastapi
uvicorn
ffmpeg
python-multipart
aiofiles
